wallet_service/requirements.txt
flask
flask-sqlalchemy
flask-migrate
gunicorn
psycopg2-binary
python-dotenv
flask-cors
